Data.IntervalMap.Interval

Description

A conservative implementation of Intervals, mostly for use as keys in a IntervalMap.

This should really be a typeclass, so you could have a tuple be an instance of Interval, but that is currently not possible in standard Haskell.

The contructor names of the half-open intervals seem somewhat clumsy, and I'm open to suggestions for better names.

Interval type

data Interval a

Intervals with endpoints of type a.

Read and Show use mathematical notation with square brackets for closed and parens for open intervals. This is better for human readability, but is not a valid Haskell expression. Closed intervals look like a list, open intervals look like a tuple, and half-open intervals look like mismatched parens.

Constructors

IntervalCO !a !a
Including lower bound, excluding upper

ClosedInterval !a !a
Closed at both ends

OpenInterval !a !a
Open at both ends

IntervalOC !a !a
Excluding lower bound, including upper

Query

lowerBound :: Interval a -> a

Get the lower bound.

upperBound :: Interval a -> a

Get the upper bound.

leftClosed :: Interval a -> Bool

Does the interval include its lower bound?

rightClosed :: Interval a -> Bool

Does the interval include its upper bound?

isEmpty :: Ord a => Interval a -> Bool

Is the interval empty?

Interval operations

overlaps :: Ord a => Interval a -> Interval a -> Bool

Do the two intervals overlap?

subsumes :: Ord a => Interval a -> Interval a -> Bool

Does the first interval completely contain the second?

before :: Ord a => Interval a -> Interval a -> Bool

Interval strictly before another? True if the upper bound of the first interval is below the lower bound of the second.

after :: Ord a => Interval a -> Interval a -> Bool

Interval strictly after another? Same as 'flip before'.

compareByUpper :: Ord a => Interval a -> Interval a -> Ordering

Like compare, but considering the upper bound first.

combine :: Ord a => Interval a -> Interval a -> Maybe (Interval a)

If the intervals overlap combine them into one.

Point operations

below :: Ord a => a -> Interval a -> Bool

Is a point strictly less than lower bound?

inside :: Ord a => a -> Interval a -> Bool

Does the interval contain a given point?

above :: Ord a => a -> Interval a -> Bool

Is a point strictly greater than upper bound?